feat: 선호 우선순위를 기반으로 룸메 추천 기능 구현

+++ src/main/java/dormitoryfamily/doomz/domain/roomate/entity/Candidate.java
@@ -28,10 +28,10 @@ public class Candidate {
     @JoinColumn(name = "candidate_member_id")
     private Member candidateMember;
 
-    private Integer candidateScore;
+    private Double candidateScore;
 
     @Builder
-    public Candidate(Recommendation recommendation, Member candidateMember, Integer candidateScore) {
+    public Candidate(Recommendation recommendation, Member candidateMember, Double candidateScore) {
         this.recommendation = recommendation;
         this.candidateMember = candidateMember;
         this.candidateScore = candidateScore;
